requirements.txt
alembic==1.0.11
aniso8601==7.0.0
attrs==19.1.0
bcrypt==3.1.7
cffi==1.12.3
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Migrate==2.5.2
flask-restplus==0.12.1
Flask-Script==2.0.6
Flask-Session==0.3.1
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==3.0.1
Mako==1.0.14
MarkupSafe==1.1.1
passlib==1.7.1
pycparser==2.19
pyrsistent==0.15.4
python-editor==1.0.4
pytz==2019.1
six==1.12.0
Werkzeug==0.15.5
requests==2.22.0
ipdb==0.12.2
gunicorn==19.9.0
Pillow==6.2.1
